refactor: add initialize in Engine for PhysicsManager
